fix(Jaeger): Externalizing Jaeger instantiation
